feat (index, style, images) add overview card
I have added the image of the back of the cards to the Overview section, formatting it to display on the left side with the descriptive text aligned to the right side in desktop view. For mobile screens, the layout adapts so that the card image appears above the text.